Pretty pointless. Save a few bytes of memory.
git-svn-id: https://fluks.googlecode.com/svn@143 6b01645e-0d02-11de-a42d-873b86d837ff

+++ tiger.h
@@ -34,7 +34,7 @@ struct tiger_ctx {
 	uint64_t	buf[8];
 	uint64_t	res[3];
 	uint64_t	length;
-	size_t		sz;
+	uint8_t		sz;
 	uint8_t		version;
 };
